Remove original component tree graph to visx
