+++ package.json
@@ -70,7 +70,6 @@
     "jsondiffpatch": "^0.4.1",
     "moment": "^2.29.1",
     "react": "^17.0.1",
-    "react-d3-tree": "^2.0.1",
     "react-dom": "^17.0.1",
     "react-fiber-traverse": "^0.0.8",
     "react-hot-loader": "^4.13.0",
